components/realtime/requirements.txt
tornado
sockjs-tornado
tornado-redis
